--- data-structures/evaluation-article/src/main/java/App.java
@@ -1,5 +1,0 @@
-public class App {
-    public static void main(String[] args) {
-        System.out.println("Hello, World!");
-    }
-}
