--- data-structures/evaluation-article/src/main/java/models/Person.java
@@ -1,32 +1,0 @@
-package models;
-
-/**
- * Represents a person with a name and age.
- */
-public class Person {
-
-    private String name;
-    private int age;
-
-    public Person(String name, int age) {
-        this.name = name;
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-}
-
