+++ evaluation-article/src/main/java/App.java
@@ -1,0 +1,5 @@
+public class App {
+    public static void main(String[] args) {
+        System.out.println("Hello, World!");
+    }
+}
