+++ evaluation-article/src/main/java/utils/DeepCopyUtil.java
@@ -1,0 +1,20 @@
+package utils;
+
+import java.io.*;
+
+public class DeepCopyUtil {
+    private DeepCopyUtil() {
+    }
+
+    public static <T extends Serializable> T deepCopy(T object) {
+        try {
+            ByteArrayOutputStream baos = new ByteArrayOutputStream();
+            ObjectOutputStream oos = new ObjectOutputStream(baos);
+            oos.writeObject(object);
+            ObjectInputStream ois = new ObjectInputStream(new ByteArrayInputStream(baos.toByteArray()));
+            return (T) ois.readObject();
+        } catch (IOException | ClassNotFoundException e) {
+            throw new IllegalArgumentException(e);
+        }
+    }
+}
